OneDrive/Desktop/New folder/internalapp/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.14
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
docopt==0.6.2
gunicorn==20.1.0
html5lib==1.1
num2words==0.5.10
Pillow==9.2.0
psycopg2==2.9.3
PyPDF2==2.6.0
python-decouple==3.6
pytz==2022.1
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.4
